Renamed duplicate fixture game_connection to mock_game_connection (this one is just a mock object) Merge/move duplicate mock_game_connection, add_connected_player and add_connected_players into unit_tests/conftest.py and reuse them
