+++ tests/unit_tests/conftest.py
@@ -1,8 +1,8 @@
-import asyncio
 from unittest import mock
 import pytest
 
 from server import GameStatsService, LobbyConnection
+from server.games import Game
 from server.gameconnection import GameConnection, GameConnectionState
 from tests import CoroMock
 
@@ -32,6 +32,14 @@ def game_connection(request, game, loop, player_service, players, game_service, 
     return conn
 
 
+@pytest.fixture
+def mock_game_connection(state=GameConnectionState.INITIALIZING, player=None):
+    gc = mock.create_autospec(spec=GameConnection)
+    gc.state = state
+    gc.player = player
+    return gc
+
+
 @pytest.fixture()
 def game_stats_service():
     service = mock.Mock(spec=GameStatsService)
@@ -59,3 +67,29 @@ def connections(loop, player_service, game_service, transport, game):
     return mock.Mock(
         make_connection=make_connection
     )
+
+def add_connected_player(game: Game, player):
+    game.game_service.player_service[player.id] = player
+    gc = mock_game_connection(state=GameConnectionState.CONNECTED_TO_HOST, player=player)
+    game.set_player_option(player.id, 'Army', 0)
+    game.set_player_option(player.id, 'StartSpot', 0)
+    game.set_player_option(player.id, 'Team', 0)
+    game.set_player_option(player.id, 'Faction', 0)
+    game.set_player_option(player.id, 'Color', 0)
+    game.add_game_connection(gc)
+    return gc
+
+
+def add_connected_players(game: Game, players):
+    """
+    Utility to add players with army and StartSpot indexed by a list
+    """
+    for army, player in enumerate(players):
+        add_connected_player(game, player)
+        game.set_player_option(player.id, 'Army', army)
+        game.set_player_option(player.id, 'StartSpot', army)
+        game.set_player_option(player.id, 'Team', army)
+        game.set_player_option(player.id, 'Faction', 0)
+        game.set_player_option(player.id, 'Color', 0)
+    game.host = players[0]
+
